warden: Modify upstart definitions
This changes makes upstart run the jobs we need in a more predictable
order. Because we use console.log to probe whether SSH has already
started or not, we need to postpone starting rsyslog or it would capture
everything written to /dev/console to /var/log/syslog inside the
container. In addition, we want SSH to start when the virtual ethernet
interface has been configured (or it still wouldn't be reachable).

Change-Id: I20ec13e30eefa5e896797f5be17680ccacdbc0a4