Fix race in stop tests
The use of iomux-spawn/iomux-link introduces a small amount of extra delay
between when spawn returns and the process is active inside the container.
This caused an infrequent race between the stop script being executed inside
the container and the spawned process being visible via /proc. This eliminates
the race by waiting for the first byte of output (or process exit) before
executing the stop.

Change-Id: Iada0bcc5b40425808a77c8925e92d6c7f21e2291