Make handles unique across restarts
A container can be identified by any string. Because this string is used
for identifying the container for kernel level resources it allocates,
the string must follow length requirements set by the kernel. This
currently comes down to a hard limit of 11 characters.

The string should only contain readable characters, preferably in the
ranges [0-9] and [a-z]. This is achieved by converting a 55 bit integer
to a string in base 32.

Uniqueness of this string is achieved by seeding the warden process with
the current time in microseconds since the Unix epoch at start, and
generate every identifier by incrementing this number.

There are two possible scenarios where the identifier generated by this
method risks not being unique:

* Warden is restarted after creating 1 or more container per microsecond
  (not possible within reason).
* Warden is restarted after the clock jumping back in time.

The second risk is a real one, and should be circumvented by having ntp
or any other time synchronization process slow the clock down instead of
making a sudden jump back in time.

Change-Id: If544aad2bdadbc3d65b39ad444b48b683896ebc7