Forward resource limits to wshd
With this change wshc reads resource limits from the environment and
passes them to wshd, that applies them before executing a process.

For every resource limit that is not specified, wshc falls back on
a set of defaults. This is done to make sure that processes that are
executed by wshd cannot inherit its resource limits.

Change-Id: Ib8b6c194d1523c3f36b72bac158ce94aa9fca960