Fix bug in special casing run command.
Copying of fields from a run request to a spawn request was earlier
assuming that all fields in the run request responded to the 'dup' call
when copying. This was wrong, for example it failed for the 'privileged'
field in run request which is of type boolean. The right way to do
this was to clone the run request through Beefcake's encode/decode
before copying its fields to the spawn request.

Change-Id: I5c1b600041c4655743ba08f9c3586efa5867aa9c