Hardcode quota blocksize
The quota blocksize is different from the filesystem blocksize. The code
originally used the filesystem blocksize to translate bytes to blocks
and blocks to bytes, which is incorrect when the quota blocksize is
different from the filesystem blocksize.

The quota blocksize is hardcoded in <sys/mount.h> as the BLOCK_SIZE
macro and is equal to 1024 by default.

Change-Id: I732081511fee503612c6c5d27cd971a1c9f0bdb7