Fix NSS initialization bug
NSS is initialized on the first call to a function defined in <pwd.h>,
such as getpwnam, getpwuid, getpwent, etc.

If this initialization step happens _after_ pivot_root(2) the shared
libraries that are picked up may be different from the ones that the
already loaded libc expects, causing inexplicable crashes.

To prevent a shared library incompatibility, trigger initialization of
the NSS before modifying the environment.

Change-Id: Ic32d7c6910c09d287091f1ff3a5e6db1243897a1