Add set that represents the original pool contents
This allows us to check whether or not an element that is released was
originally contained in the pool or not. It is possible that a pool
configuration is changed, and that elements that were acquired from the
original pool are contained in the new pool. When such elements are
released, they should be ignored.

Change-Id: I9063d9148d8018e7324bd152b31f91106ffa5160