Always recover container snapshots
This used to only take place when the warden server was properly
drained. In reality, it is more likely that is is NOT properly drained
and just gets killed.

This commit doesn't cover every edge case but it a first attempt in
making this happen. Things that need to be checked/put in place are:

- The OOM notifier must run after recovery
- Memory limits must be persisted to and recovered from the snapshot
- Bandwidth limits must be persisted to and recovered from the snapshot
- The disk quota must be persisted to and recovered from the snapshot

Change-Id: I0d4fed1af58723d5f33c6be107f0cc2db999db19