add log_tag to spawn/run requests
This is used to direct the stdout/stderr of the job to some syslog-like
service with the given tag.

Also redo protocol building.

A previous commit added explicit imports to the separate .protos so they
could be built independently, but we were building it all by appending
them to one file, so it had duplicate imports. This builds them
separately and appends all of the output to the single .rb afterwards.

Signed-off-by: Maria Shaldibina <mariash@pivotallabs.com>