Sync fs and move image deletion after loop detach
The warden tests frequently fail under VMware due to an error detatching
a loop device that was associated with a mounted disk image:

  loop: can't delete device /dev/loop0: Device or resource busy

In order to reduce the likelihood of the problem, a `sync` was added
before the detach and the deletion of the image file was moved until
after the detach.  After repeated runs with these modifications, I have
not seen the failures occur.